tests/requirements.txt
pytest==8.3.2
pytest-playwright==0.5.5
